Remove deprecated features
